+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Mind-blowingly simple local resource management
 
 ##ChangeLog
 
+### Development HEAD
+- Add priority option to allow semaphore requests to jump in line
+
 ### v0.1.1
 - Drop attempted (but already non-functioning) support for pre-0.10.0 version
 of node; process.nextTick is too unpredictable